Move some shared definitions back to src/main/scala/
